user input updating redux store, which is updating username fetch which is fetching as intended
