+++ src/components/welcome-component/UserNameFetch.jsx
@@ -1,10 +1,15 @@
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
+import { useSelector, useDispatch } from "react-redux";
+import { selectUserName } from "../userSlice/userSlice";
+import { updateUserName } from "../userSlice/userSlice";
 
 const UserNameFetch = () => {
-  const queryParams = "Benny";
+  const userName = useSelector(selectUserName);
+  console.log("userName in usernamefetch", userName);
+  const queryParams = userName;
   const url = `http://localhost:8888/getusername?enteredName=${queryParams}`;
-  const [name, setName] = useState("default name in state");
-  console.log("name in state", name);
+
+  const dispatch = useDispatch();
 
   // fetch user name from user input, if name does not exist add name to server, create new user object
   useEffect(() => {
@@ -17,10 +22,11 @@ const UserNameFetch = () => {
       })
       .then((data) => {
         console.log("data", data);
-        setName(data.userName);
+
+        dispatch(updateUserName(data.userName));
         console.log(name);
       });
-  }, []);
+  }, [userName]);
 
   return <div>{name}</div>;
 };
