+++ src/store.js
@@ -1,8 +1,8 @@
 import { configureStore } from "@reduxjs/toolkit";
-import nameSlice from "./components/userName/nameSlice";
+import userSlice from "./components/userSlice/userSlice";
 
 export const store = configureStore({
   reducer: {
-    nameSlice: nameSlice,
+    userSlice: userSlice,
   },
 });
